pyproject.toml
[project]
name = "quant-copilot"
version = "0.1.0"
description = ""
authors = [
    {name = "Mak-1911",email = "asif3702@gmail.com"}
]
readme = "README.md"
packages = [{ include = "app" }]
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "openai (>=1.96.1,<2.0.0)",
    "backtrader (>=1.9.78.123,<2.0.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "yfinance (>=0.2.65,<0.3.0)",
    "plotly (>=6.2.0,<7.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "transformers (>=4.53.2,<5.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "google-generativeai (>=0.7.0,<0.8.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "aiosmtplib (>=4.0.1,<5.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
